Fix broken GitHub workflow
The version_changed artifact from the check job wasn't being properly
propagated to the second, publish, job, resulting in the publish job
always being skipped. This should fix that problem. More here:

https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/passing-information-between-jobs

(we bump the version number in this commit to test that our change
works)

Also, we need to compare the current version in package.json, not to the
latest version in the latest commit on the main branch (which will
always be the same), but to the version in the second to last commit on
main to decide whether or not to publish to the npm registry.